seperated component into render and container
